+++ foliant/backends/confluence/convert.py
@@ -2,6 +2,7 @@ import re
 import os
 import shutil
 
+from filecmp import cmp
 from pathlib import PosixPath, Path
 from subprocess import run, PIPE, STDOUT
 
@@ -118,7 +119,7 @@ def process_images(source: str,
     shutil.rmtree(target_dir, ignore_errors=True)
     Path(target_dir).mkdir()
 
-    image_pattern = re.compile(r'<im§1g src="(?P<path>.+?)" +(?:alt="(?P<caption>.*?)")?.+?>')
+    image_pattern = re.compile(r'<img src="(?P<path>.+?)" +(?:alt="(?P<caption>.*?)")?.+?>')
     attachments = []
 
     logger.debug('Processing images')
@@ -194,18 +195,39 @@ def add_toc(source: str) -> str:
     return result
 
 
-def update_attachments(page: Page, attachments: list, title: str):
+def update_attachments(page: Page,
+                       attachments: list,
+                       cache_dir: PosixPath or str):
+    '''
+    Upload a list of attachments into page. Only changed attachments will
+    be updated. If page doesn't exist yet, an empty one will be created.
+
+    `page` — a Page object to which attachments will be uploaded.
+    `attachments` — a list of attachments PosixPaths.
+    `cache_dir` — temporary dir where old attachments will be downloaded to
+                  for comparison.
+    '''
     if attachments:
         # we can only upload attachments to existing page
         if not page.exists:
             logger.debug('Page does not exist. Creating an empty one '
                          'to upload attachments')
-            page.create_empty_page(title)
+            page.create_empty_page()
         else:
-            logger.debug('Removing old attachments and adding new')
-            page.delete_all_attachments()
-        for img in attachments:
-            page.upload_attachment(img)
+            cache_dir = Path(cache_dir)
+            shutil.rmtree(cache_dir, ignore_errors=True)
+            cache_dir.mkdir(exist_ok=True)
+            remote_dict = page.download_all_attachments(cache_dir)
+            for att in attachments:
+                if att.name in remote_dict:
+                    att_id, att_path = remote_dict[att.name]
+                    if cmp(att, att_path):  # attachment not changed
+                        logger.debug(f"Attachment {att.name} hadn't changed, skipping")
+                        continue
+                logger.debug(f"Attachment {att.name} CHANGED, reuploading")
+                # not sure if it's needed, we can update images without deleting
+                # page.delete_attachment(att_id)
+                page.upload_attachment(att)
 
 
 def set_up_logger(logger_):